feat(pagination): implement keyset cursor pagination for comments, notifications, and follow lists (TASK-10.8)
- Add CursorPageResponse<T> (items + nextCursor + hasMore) as standard paginated response
- Add CursorEncoder: opaque Base64 encoding/decoding of (createdAt, id) cursor pairs
- Replace OFFSET queries with native SQL row-comparison keyset queries in CommentJpaRepository, NotificationJpaRepository, and FollowJpaRepository
- Update CommentController, NotificationController, FollowController to accept cursor param and return CursorPageResponse
- Add GetFollowersUseCase.FollowersPage and GetFollowingUseCase.FollowingPage nested records so services can return both items and nextCursor
- Update frontend useNotifications, useFollowers, useFollowing hooks to use useInfiniteQuery with nextCursor instead of page-index offset
- Fix CommentService NPE when cursor is null; fix @Cacheable condition on cursor-based queries
- Update all affected tests: CommentServiceTest, LikeServiceTest, MessagingServiceTest, NotificationServiceTest, FollowServiceTest, CommentControllerTest, FollowControllerIT, NotificationControllerIT, CommentPersistenceAdapterIT, NotificationPersistenceAdapterIT (210/210 green)

+++ backend/src/main/java/com/instagram/adapter/in/web/dto/response/ConversationResponse.java
@@ -7,10 +7,10 @@ import com.instagram.domain.model.Conversation;
 
 public record ConversationResponse(
         UUID id, String name, boolean isGroup, MessageResponse lastMessage,
-        int unreadCount, String eachOtherName, OffsetDateTime createdAt) {
+        long unreadCount, String eachOtherName, OffsetDateTime createdAt) {
 
     public static ConversationResponse from(Conversation conversation, MessageResponse lastMessage,
-            int unreadCount, String eachOtherName) {
+            long unreadCount, String eachOtherName) {
         return new ConversationResponse(conversation.getId(), conversation.getName(), conversation.isGroup(),
                 lastMessage, unreadCount, eachOtherName, conversation.getCreatedAt());
     }
